pushing project 3
This project explores salary trends and residency patterns in the field of data science through linear and logistic regression analysis. Using a global dataset of data science professionals, it examines the impact of expertise level, location, and salary on trends and predictions. The analysis includes predictive modeling, residual diagnostics, and classification metrics to uncover key insights and address research questions.